chore: rename prefabeditors to componenteditors and remove advanced mode add button
