+++ packages/editor/src/components/element/ElementList.tsx
@@ -65,8 +65,8 @@ import { GroupAddOutlined as PlaceHolderIcon } from '@mui/icons-material'
 import { IconButton, PopoverPosition } from '@mui/material'
 
 import { ItemTypes } from '../../constants/AssetTypes'
+import { EntityNodeEditor } from '../../functions/ComponentEditors'
 import { EditorControlFunctions } from '../../functions/EditorControlFunctions'
-import { EntityNodeEditor } from '../../functions/PrefabEditors'
 import { getCursorSpawnPosition, getSpawnPositionAtCenter } from '../../functions/screenSpaceFunctions'
 import { ContextMenu } from '../layout/ContextMenu'
 import styles from './styles.module.scss'
