+++ packages/editor/src/components/hierarchy/HierarchyTreeNode.tsx
@@ -51,9 +51,9 @@ import ArrowRightIcon from '@mui/icons-material/ArrowRight'
 
 import { ItemTypes, SupportedFileTypes } from '../../constants/AssetTypes'
 import { addMediaNode } from '../../functions/addMediaNode'
+import { EntityNodeEditor } from '../../functions/ComponentEditors'
 import { EditorControlFunctions } from '../../functions/EditorControlFunctions'
 import { isAncestor } from '../../functions/getDetachedObjectsRoots'
-import { EntityNodeEditor } from '../../functions/PrefabEditors'
 import { SelectionState } from '../../services/SelectionServices'
 import useUpload from '../assets/useUpload'
 import { addSceneComponentElement } from '../element/ElementList'
