+++ packages/editor/src/components/properties/CoreNodeEditor.tsx
@@ -38,15 +38,11 @@ import {
 import { EntityOrObjectUUID, getEntityNodeArrayFromEntities } from '@etherealengine/engine/src/ecs/functions/EntityTree'
 import { SceneTagComponent } from '@etherealengine/engine/src/scene/components/SceneTagComponent'
 import { VisibleComponent } from '@etherealengine/engine/src/scene/components/VisibleComponent'
-import { dispatchAction, getMutableState, useHookstate } from '@etherealengine/hyperflux'
-
-import AddIcon from '@mui/icons-material/Add'
+import { getMutableState, useHookstate } from '@etherealengine/hyperflux'
 
 import { EditorControlFunctions } from '../../functions/EditorControlFunctions'
-import { EntityNodeEditor } from '../../functions/PrefabEditors'
 import { EditorState } from '../../services/EditorServices'
-import { SelectionAction, SelectionState } from '../../services/SelectionServices'
-import MainMenu from '../dropDownMenu'
+import { SelectionState } from '../../services/SelectionServices'
 import BooleanInput from '../inputs/BooleanInput'
 import InputGroup from '../inputs/InputGroup'
 import NameInputGroup from './NameInputGroup'
@@ -107,23 +103,6 @@ export const CoreNodeEditor: EditorComponentType = (props) => {
           </>
         )}
       </NameInputGroupContainer>
-
-      {/** @todo #7258 */}
-      {editorState.advancedMode.value && typeof props.entity === 'number' && (
-        <div style={{ pointerEvents: 'auto' }}>
-          <MainMenu
-            icon={AddIcon}
-            commands={Array.from(EntityNodeEditor).map(([component, editor]) => ({
-              name: component.name,
-              action: () => {
-                if (!ComponentMap.get(component.name)) return console.warn('could not find component', component.name)
-                setComponent(props.entity, ComponentMap.get(component.name)!)
-                dispatchAction(SelectionAction.forceUpdate({}))
-              }
-            }))}
-          />
-        </div>
-      )}
     </PropertiesHeader>
   )
 }
